Relocated API project and renamed "service" to "api"
To mirror the internal project structure, so as to make it easier to replicate for in-box build, the WinRT API was moved to the Client folder under what used to be titled "service". And that folder was itself renamed to "api" as the API is the service, transports, WinRT API etc. all together. That's my story and I'm sticking to it.