Installer working, despite its very best efforts not to
I dragged it kicking and screaming into a working state. Gave up on trying to find command switches / properties to force it to build 64 bit and just reordered the targets in the .sln so it would do it. Also, the registry functions are now in place, and were much easier to do 100% in C#, so there's a step in the cake build script which now builds are array of objects that gets included in the custom WiX action .NET dll at build time.