Fix problem where not all WinMM endpoints would be cleaned up
The service adds the group index to the end of the WinMM device instance id, so there was not an exact match when a transport would disconnect the device, and the WinMM ports wouldn't always be cleaned up (in the Virtual Device transport, for example). Changed to fix that and revert back a couple other changes I made while debugging this.